Build with Netlify
Mainly because of `baseURL` issues, I cannot deploy this site properly
with its assets. To fix it, I just have to properly set them and the
easiest way to do so is to build it with Netlify.

With this change, the webring post update will be in a separate workflow
and done in a schedule.